+++ examples/full/config.js
@@ -1,7 +1,7 @@
 import routes from 'examples/full/routes';
 import webpackConfig from 'examples/full/webpack.config';
 
-const enableServerRender = false;
+const enableServerRender = true;
 const enableClientRender = true;
 const enableReduxSimpleRouter = true;
 const enableDevTools = true;